Added Container Mixin, updated Body/Room, redid Body addressing.
Ugh, Body location.. it now has a location STRING and the room object.
I'm not sure I like this yet, but it seems to work.  I did it to get rid
of the issue of saving EVERYTHING IN THE ROOM EVAR when I try to persist
a player.  Added a method GetRoom() to body to return the room object.

I broke all the pickled data too.  Goddamnit.