Raising a value error if you can't find the room is a bad plan, for now.
I think I'd have to rearchitect the whole system if I wanted to do this.
Exceptions do have advantages in that I can get Tracebacks, which help
debugging.  But if I do that here, I need to do it EVERYWHERE... instead
of returning None, throw an error.. then I have to make sure those are
handled EVERYWHERE.  Returning None from stuff is my way of saying "this
didn't work"... which means I can launch Nones down a chain.  Though
Exceptions I think are the "right" way to do it, I'm going to stick with
the way I have it.