Very early beginnings of a parser.
Well, first we need to start operating out of a parser object.  I'm using this as a singleton right now, which seems to work, but I'm wondering if I shouldn't spawn one for each incoming message?  I don't think that's a good idea.  I think I'd like to pre-load my verbs, so having just one Parser floating around that has all the verbs in it will be a Good Thing.

Also added a framework for a Verb class.  Haven't decided how to actually CODE the verbs, but we'll figure it out.